* commit '621d8785de13aa693350596f967dfcdba9d135a6': (664 commits)
  removed versionchanged as no backwards compat stuff happened
  Add etcd update function
  fix get_enabled to return currently enable services
  fix typo
  expect the new exception that can be thrown and handle it
  switch unit test to except CommandExecutionError
  switch to raising an exception as mentioned in #30578
  Fix test failures
  Fix salt.states.archive __virtual__
  put back the call to subshell `sh -c`
  Updated release notes for boron to reflect filename changes
  Merge #30610 with pylint fix
  move password_column to the end of kwargs
  Updates made after PR review:
  Swapping one unknow for debian failed, because debian use a different function internally. Switched to RedHat for the test as it returns LANG=XYZ
  Put in pylint forward compatibility recommendation.
  update tests, we have a correct return now
  Update mysql.py
  Retry waiting for IAM role to be propagated
  fixed
  ...

+++ salt/modules/parted.py
@@ -395,7 +395,11 @@ def mkfs(device, fs_type):
                           'hfs', 'hfs+', 'hfsx', 'NTFS', 'ufs']):
         raise CommandExecutionError('Invalid fs_type passed to partition.mkfs')
 
-    mkfs_cmd = 'mkfs.{0}'.format(fs_type)
+    if fs_type is 'linux-swap':
+        mkfs_cmd = 'mkswap'
+    else:
+        mkfs_cmd = 'mkfs.{0}'.format(fs_type)
+
     if not salt.utils.which(mkfs_cmd):
         return 'Error: {0} is unavailable.'
     cmd = '{0} {1}'.format(mkfs_cmd, device)
@@ -640,6 +644,9 @@ def set_(device, minor, flag, state):
         "on" or "off". Some or all of these flags will be available, depending
         on what disk label you are using.
 
+    Valid flags are: bios_grub, legacy_boot, boot, lba, root, swap, hidden, raid,
+        LVM, PALO, PREP, DIAG
+
     CLI Example:
 
     .. code-block:: bash
@@ -671,13 +678,14 @@ def toggle(device, partition, flag):
     '''
     partition.toggle device partition flag
 
-    Toggle the state of <flag> on <partition>
+    Toggle the state of <flag> on <partition>. Valid flags are the same as
+        the set command.
 
     CLI Example:
 
     .. code-block:: bash
 
-        salt '*' partition.name /dev/sda 1 boot
+        salt '*' partition.toggle /dev/sda 1 boot
     '''
     _validate_device(device)
